pyproject.toml
[project]
name = "bankiru-parser"
version = "0.1.0"
description = "Banki.ru negative reviews parser"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "emoji>=2.14.1",
    "environs>=14.3.0",
    "logfire>=4.3.3",
    "pandas>=2.3.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "schedule>=1.2.2",
]
